+++ RELEASES.md
@@ -3,6 +3,7 @@
 ## Release 2024-11-03
 
 - ✨ Zie sneller waar hubs staan op de kaart: Uitgezoomd zie je het [Mijksenaar hub-logo](https://www.zeeland.nl/sites/default/files/2022-07/2022-07-11%20Handboek%20identiteit%20hubs%20V1.pdf), ingezoomd de hubcontouren zelf
+- ✨ Bekijk de datafeed-status van elke aanbieder in [dit statusoverzicht](https://dashboarddeelmobiliteit.nl/active_feeds)
 
 ## Release 2024-11-01